% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_file_extension.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Get File Extension}
\usage{
get_file_extension(file_name)
}
\arguments{
\item{file_name}{The name of the uploaded file.}
}
\value{
A string representing the file extension (e.g., "csv", "txt").
}
\description{
A function to extract the file extension from a file name.
}
