% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_list_of_variables.R
\name{get_list_of_variables}
\alias{get_list_of_variables}
\title{Get List of Variables (Column Names) from Uploaded Data}
\usage{
get_list_of_variables(data)
}
\arguments{
\item{data}{A data frame from which to extract column names.}
}
\value{
A vector containing the column names.
}
\description{
A function to extract the column names from a data frame.
}
